fix(pack,bytes): truncate (not saturate) high-bit values in pack/chr
Phase B of dev/modules/jcpan_bundled_dbd_and_convert_ber.md.
Two bugs surfaced while triaging Convert::BER under jperl
(prereq for DBD::JDBC):

1. pack 'c'/'C'/'s'/'S'/'n'/'v' saturated for doubles whose value
   exceeds Integer.MAX_VALUE, because the call sites read
   value.getInt() which goes (int)(double). For example
   pack('n', 0x81828384) yielded "\xff\xff" instead of the correct
   "\x83\x84" (low 16 bits, matching system Perl and what 'N' / 'V'
   already do).
   Fix: route the wide formats through (int)(long)value.getDouble(),
   so we truncate the long bit pattern and let the writeShort* /
   output.write masking pick the low N bits. NumericPackHandler.java.

2. Under `use bytes`, ord(chr(0x84)) yielded 0xc2 (UTF-8 first byte of
   codepoint 0x84) instead of 0x84. Two sub-bugs:
   - chrBytes used getInt() and inherited the same saturation bug;
     for chr(0x81828384) we got 0xff & 0xFF instead of 0x84.
   - ordBytes always called getBytes(UTF_8) on the input String and
     returned bytes[0]; but BYTE_STRING already stores raw bytes
     (each Java char is a 0-255 byte value), so re-encoding to UTF-8
     was wrong.
   Fix: chrBytes goes through (long)getDouble() & 0xFF; ordBytes
   takes a BYTE_STRING fast path that returns charAt(0) & 0xFF.
   StringOperators.java, ScalarOperators.java.

Adds a regression subtest in src/test/resources/unit/pack.t covering
the high-bit values for c/C/s/n/v/nc and the chr/ord round-trip.

Effect on Convert-BER-1.32 t/00prim.t: 36 -> 33 failures. Encode is
now correct; decode failures are deeper and are tracked in the doc
under "Remaining" along with a separate float32-rounding bug found
on the numeric->string path (printf "%d", 0x81828384 prints
2172814212).

+++ src/main/java/org/perlonjava/runtime/operators/pack/NumericPackHandler.java
@@ -128,28 +128,28 @@ public class NumericPackHandler implements PackFormatHandler {
             switch (format) {
                 case 'c':
                     // Signed char
-                    int signedChar = value.getInt();
+                    int signedChar = (int) (long) value.getDouble();
                     output.write(signedChar & 0xFF);
                     break;
                 case 'C':
                     // Unsigned char
-                    int intValue = value.getInt();
+                    int intValue = (int) (long) value.getDouble();
                     output.write(intValue & 0xFF);
                     break;
                 case 's':
                     // Signed short - use endianness if specified
                     if (modifiers.bigEndian) {
-                        PackWriter.writeShortBigEndian(output, value.getInt());
+                        PackWriter.writeShortBigEndian(output, (int) (long) value.getDouble());
                     } else {
-                        PackWriter.writeShortLittleEndian(output, value.getInt());
+                        PackWriter.writeShortLittleEndian(output, (int) (long) value.getDouble());
                     }
                     break;
                 case 'S':
                     // Unsigned short - use endianness if specified
                     if (modifiers.bigEndian) {
-                        PackWriter.writeShortBigEndian(output, value.getInt());
+                        PackWriter.writeShortBigEndian(output, (int) (long) value.getDouble());
                     } else {
-                        PackWriter.writeShort(output, value.getInt());
+                        PackWriter.writeShort(output, (int) (long) value.getDouble());
                     }
                     break;
                 case 'l':
@@ -199,7 +199,7 @@ public class NumericPackHandler implements PackFormatHandler {
                     break;
                 case 'n':
                     // Network short (always big-endian)
-                    PackWriter.writeShortBigEndian(output, value.getInt());
+                    PackWriter.writeShortBigEndian(output, (int) (long) value.getDouble());
                     break;
                 case 'N':
                     // Network long (always big-endian)
@@ -207,7 +207,7 @@ public class NumericPackHandler implements PackFormatHandler {
                     break;
                 case 'v':
                     // VAX short (always little-endian)
-                    PackWriter.writeShortLittleEndian(output, value.getInt());
+                    PackWriter.writeShortLittleEndian(output, (int) (long) value.getDouble());
                     break;
                 case 'V':
                     // VAX long (always little-endian)
